On the listings/[id] page, if a business is not an in-house business (i.e. it's a Google listing), we now: - Hide the tabs for promotion, product, service, etc. - Show a "Claim this listing" button and a "Share" button. - Show the business overview below the buttons.
This was achieved by:
- Refactoring the `OverviewSection` component out of `ContentTabs.tsx` into its own file (`components/OverviewSection.tsx`).
- In `ListingPage.tsx`, using the `isGoogle` prop to conditionally render either the `ContentTabs` component (for in-house listings) or the new layout with the buttons and the `OverviewSection` (for Google listings).
