+++ components/ContentTabs.tsx
@@ -262,249 +262,7 @@ function ServicePage({
   );
 }
 
-function OverviewSection({
-  listing,
-  isLoading,
-}: {
-  listing: GooglePlaceResult | InHouseBusiness;
-  isLoading: boolean;
-}) {
-  const [isCopied, setIsCopied] = useState(false);
-  const router = useRouter();
-
-  const isGoogle = isGoogleResult(listing);
-  const today = new Date().getDay();
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(window.location.href).then(
-      () => {
-        setIsCopied(true);
-        setTimeout(() => setIsCopied(false), 2000); // Reset after 2 seconds
-      },
-      err => {
-        console.error('Failed to copy: ', err);
-      }
-    );
-  };
-
-  const formatTime = (timeString: string) => {
-    if (!timeString) return '';
-    const [hours, minutes] = timeString.split(':');
-    let h = parseInt(hours, 10);
-    const ampm = h >= 12 ? 'PM' : 'AM';
-    h = h % 12;
-    h = h ? h : 12; // the hour '0' should be '12'
-    return `${h}:${minutes} ${ampm}`;
-  };
-
-  const daysOfWeek = [
-    'Sunday',
-    'Monday',
-    'Tuesday',
-    'Wednesday',
-    'Thursday',
-    'Friday',
-    'Saturday',
-  ];
-
-  const location = isGoogle ? listing.geometry : listing.location;
-  const address = isGoogle
-    ? listing.formattedAddress || listing.vicinity
-    : `${listing.location.addressLine1}, ${listing.location.city}`;
-  const reviews = isGoogle ? listing.reviews : []; // In-house doesn't have reviews yet
-  const businessId = isGoogle
-    ? (listing as GooglePlaceResult).placeId
-    : (listing as InHouseBusiness).id;
-
-  if (isGoogle) {
-    // Keeping Google result view simpler as requested
-    return (
-      <div className="space-y-6">
-        <div className="bg-gray-50 p-6 rounded-lg">
-          <h3 className="text-xl font-bold text-gray-800 mb-4">
-            Business Details
-          </h3>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4 text-gray-700">
-            <p>
-              <strong>Status:</strong> {listing.businessStatus}
-            </p>
-            <p>
-              <strong>Types:</strong> {listing.types?.join(', ')}
-            </p>
-            {listing.openingHours && (
-              <p>
-                <strong>Availability:</strong>{' '}
-                {listing.openingHours.openNow ? 'Open Now' : 'Closed'}
-              </p>
-            )}
-          </div>
-        </div>
-        <div className="py-8">
-          <LocationSection listing={location} address={address} />
-        </div>
-        <div className="bg-gray-50 py-8 px-6 rounded-lg">
-          <h3 className="text-2xl font-bold text-gray-800 mb-4">FAQ</h3>
-          <p className="text-gray-600">FAQ content goes here.</p>
-        </div>
-        <div className="py-8">
-          <ReviewsTabContent businessId={businessId} />
-        </div>
-      </div>
-    );
-  }
-
-  // InHouseBusiness
-  return (
-    <div className="-mx-6">
-      {!isGoogle && !(listing as InHouseBusiness).isClaimed && (
-        <div className="bg-yellow-50 border-l-4 border-yellow-400 p-6 mx-6 mb-8 rounded-r-lg">
-          <h3 className="text-xl font-bold text-yellow-800">
-            Do you know the owner of this business?
-          </h3>
-          <p className="text-yellow-700 mt-2">
-            Tell them to claim this listing to unlock more features and manage
-            their business information.
-          </p>
-          <Button
-            onClick={handleCopy}
-            className="mt-4 bg-yellow-500 hover:bg-yellow-600 text-white"
-          >
-            {isCopied ? 'Copied!' : 'Copy Listing URL'}
-          </Button>
-        </div>
-      )}
-
-      {/* About Section */}
-      <div className="py-8 px-6">
-        <h3 className="text-2xl font-bold text-gray-800 mb-3">
-          About {listing.businessName}
-        </h3>
-        <p className="text-gray-600 leading-relaxed">
-          {listing.about || listing.shortDescription}
-        </p>
-      </div>
-
-      {/* Opening Hours Section */}
-      {listing.businessHours && listing.businessHours.length > 0 && (
-        <div className="bg-slate-50 py-8 px-6">
-          <h3 className="text-2xl font-bold text-gray-800 mb-4">
-            Opening Hours
-          </h3>
-          <ul className="space-y-2">
-            {listing.businessHours
-              .sort((a, b) => a.dayOfWeek - b.dayOfWeek)
-              .map(hour => (
-                <li
-                  key={hour.id}
-                  className={`flex justify-between p-3 rounded-lg ${
-                    hour.dayOfWeek === today
-                      ? 'bg-red-100 text-red-800'
-                      : 'text-gray-700'
-                  }`}
-                >
-                  <span className="font-semibold">
-                    {daysOfWeek[hour.dayOfWeek]}
-                  </span>
-                  <span
-                    className={
-                      hour.dayOfWeek === today ? 'font-bold' : ''
-                    }
-                  >
-                    {hour.is24h
-                      ? '24 Hours'
-                      : `${formatTime(hour.openTime)} - ${formatTime(
-                          hour.closeTime
-                        )}`}
-                  </span>
-                </li>
-              ))}
-          </ul>
-        </div>
-      )}
-
-      {/* Amenities Section */}
-      {(listing.productSellerProfile || listing.serviceProviderProfile) && (
-        <div className="py-8 px-6">
-          <h3 className="text-2xl font-bold text-gray-800 mb-4">
-            Amenities & Services
-          </h3>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-            {listing.productSellerProfile && (
-              <>
-                <div className="flex items-start space-x-3">
-                  <Truck className="h-6 w-6 text-red-500 flex-shrink-0 mt-1" />
-                  <div>
-                    <h4 className="font-semibold">Selling Modes</h4>
-                    <p className="text-gray-600">
-                      {listing.productSellerProfile.sellingModes.join(', ')}
-                    </p>
-                  </div>
-                </div>
-                {listing.productSellerProfile.returnsPolicy && (
-                  <div className="flex items-start space-x-3">
-                    <Info className="h-6 w-6 text-red-500 flex-shrink-0 mt-1" />
-                    <div>
-                      <h4 className="font-semibold">Returns Policy</h4>
-                      <p className="text-gray-600">
-                        {listing.productSellerProfile.returnsPolicy}
-                      </p>
-                    </div>
-                  </div>
-                )}
-              </>
-            )}
-            {listing.serviceProviderProfile && (
-              <div className="flex items-start space-x-3">
-                <Clock className="h-6 w-6 text-red-500 flex-shrink-0 mt-1" />
-                <div>
-                  <h4 className="font-semibold">Booking Method</h4>
-                  <p className="text-gray-600">
-                    {listing.serviceProviderProfile.bookingMethod}
-                  </p>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-
-      {/* Contact Info Section */}
-      {(listing.website || listing.businessEmail) && (
-        <div className="bg-slate-50 py-8 px-6">
-          <h3 className="text-2xl font-bold text-gray-800 mb-4">
-            Contact
-          </h3>
-          <div className="space-y-3">
-            {listing.website && (
-              <a
-                href={listing.website}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="flex items-center text-red-500 hover:underline"
-              >
-                <Globe className="mr-2 h-5 w-5" />
-                {listing.website}
-              </a>
-            )}
-          </div>
-        </div>
-      )}
-
-      <div className="py-8 px-6">
-        <LocationSection listing={location} address={address} />
-      </div>
-
-      <div className="bg-slate-50 py-8 px-6">
-        <h3 className="text-2xl font-bold text-gray-800 mb-4">FAQ</h3>
-        <p className="text-gray-600">FAQ content goes here.</p>
-      </div>
-
-      <div className="py-8 px-6">
-        <ReviewsTabContent businessId={businessId} />
-      </div>
-    </div>
-  );
-}
+import OverviewSection from './OverviewSection';
 
 function PromotionsTabs({
   listing,
